refactor(common/constants): move ignoreSignerCheckBlockArray and denylist to shared global file
- Consolidate ignoreSignerCheckBlockArray and denylist into constants.shared.go for all networks
- Remove per-network duplication and related code
- Update related functions to use shared global variables
